Trading/trading-bot-a69v/requirements.txt
asyncio
pandas>=2.0.0
numpy>=1.24.0
ccxt>=4.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
websockets>=12.0
PyYAML>=6.0
requests>=2.31.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
pylint>=3.0.0
